(feature) Implement the first question
Show users the first question and two possible answers. Depending
on the chosen answer, the user is either redirected to the emergency
contacts page, or shown the next question (currently, faked in
view of `Questions::StartController#submit`).

In the case where the form is submitted without choosing an option, the
users are shown a warning message and the form is redisplayed.