(feature) Text entry questions redirect after submission
Previously, only questions with answers could cause a redirect. Now the
`goto` key can be set at the question level.

I've assumed that after text entry, you'll always be redirecting to
another question rather than static content. If `goto` is not set, it'll
default to redirecting the the address search form, which is the first
page AFTER the decision tree.